Keep merged branch within Swift file budget
After merging origin/main, the existing CLI and Workspace growth exceeded the repo file-length guard. Moving the CLI Sentry telemetry helper into its own support file keeps the large CLI entrypoint under budget without changing behavior.

Constraint: origin/main had to be merged before updating the PR; no local tests or direct xcodebuild are allowed.

Rejected: Refresh the Swift file length budget | this is branch-owned growth and can be reduced by moving an existing helper.

Confidence: medium

Scope-risk: narrow

Directive: Keep large CLI support helpers outside CLI/cmux.swift when adding new hook or socket behavior.

Tested: git diff --check origin/main...HEAD; plutil -lint GhosttyTabs.xcodeproj/project.pbxproj; python3 scripts/swift_file_length_budget.py

Not-tested: Local XCTest/build execution intentionally skipped by instruction; CI must validate compile and tests.
